Time: 0 ms (100%), Space: 7.9 MB (29.33%) - LeetHub

+++ 0028-find-the-index-of-the-first-occurrence-in-a-string/0028-find-the-index-of-the-first-occurrence-in-a-string.cpp
@@ -28,7 +28,7 @@ public:
 
         // Finding first occurence of 'needle' in 'haystack'
         int l=0,r=0;
-        for(int i=0;i<=haystack.size()-needle.size();i++)
+        for(int i=0;i<=(int)haystack.size()-(int)needle.size();i++)
         {
             int len=0;
             if(i<r)
